Move TransactionDb
