+++ src/brs/BlockchainProcessorImpl.java
@@ -6,6 +6,7 @@ import brs.at.AT_Exception;
 import brs.crypto.Crypto;
 import brs.db.BlockDb;
 import brs.db.DerivedTable;
+import brs.db.TransactionDb;
 import brs.db.cache.DBCacheManagerImpl;
 import brs.db.store.BlockchainStore;
 import brs.db.store.DerivedTableManager;
